+++ CHANGELOG.md
@@ -19,6 +19,7 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
     - `strip_frontmatter`: If `true`, strip frontmatter from .md documents (i.e. `---` enclosed block)
     - `number_headings`: If `true`, add numbers to headings.
     - `number_headings_at_level`: If `number_headings` is true, then this controls the heading level at which numbering will begin. This defaults to `1`, but typically a user will set this to `2`if they have only one level 1 heading, and that heading is the title of the document.
+    - `menu_level_two_number_prefix_width`: (optional) Width to use for heading number field in level 2 menu headings (e.g. `20px`)
 ## Removed
 - Command line `--nobreak` option